Refactoring components into core
