+++ src/earthkit/plots/core/figures.py
@@ -19,10 +19,10 @@ import matplotlib.image as mpimg
 import matplotlib.pyplot as plt
 
 from earthkit.plots.ancillary import find_logo
-from earthkit.plots.components.layers import LayerGroup
-from earthkit.plots.components.layouts import rows_cols
-from earthkit.plots.components.maps import Map
-from earthkit.plots.components.subplots import Subplot
+from earthkit.plots.core.layers import LayerGroup
+from earthkit.plots.core.layouts import rows_cols
+from earthkit.plots.core.subplots import Subplot
+from earthkit.plots.geospatial.maps import Map
 from earthkit.plots.metadata import formatters
 from earthkit.plots.schemas import schema
 from earthkit.plots.utils import string_utils
