+++ src/earthkit/plots/geospatial/__init__.py
@@ -11,11 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-from earthkit.plots.core import subplots
-
-
-def test_subplots_figure_size():
-    # Test that the figure size is set correctly when creating a subplot
-    subplot = subplots.Subplot(size=[10, 5])
-    assert subplot.figure._figsize == [10, 5]
