+++ src/earthkit/plots/geospatial/globes.py
