+++ src/earthkit/plots/geospatial/maps.py
@@ -17,7 +17,7 @@ import cartopy.feature as cfeature
 import cartopy.io.shapereader as shpreader
 import matplotlib.patheffects as pe
 
-from earthkit.plots.components.subplots import Subplot
+from earthkit.plots.core.subplots import Subplot
 from earthkit.plots.geo import coordinate_reference_systems, domains, natural_earth
 from earthkit.plots.metadata.formatters import SourceFormatter
 from earthkit.plots.metadata.labels import CRS_NAMES
