+++ src/earthkit/plots/geospatial/tiles.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from earthkit.plots.components.maps import Map
+from earthkit.plots.geospatial.maps import Map
 
 
 class Tile(Map):
