+++ src/earthkit/plots/interactive/animations.py
